libs/resource-plugin/src/generators/app/generator.ts
import {
  formatFiles,
  generateFiles,
  joinPathFragments,
  Tree,
} from '@nx/devkit';
import { applicationGenerator } from '@nx/nest';

import { AppGeneratorSchema } from './schema';
import { paramCase, pascalCase, snakeCase } from 'change-case';

export async function appGenerator(tree: Tree, options: AppGeneratorSchema) {
  options.name = paramCase(options.name);
  const appRoot = `apps/${options.name}`;
  const srcFolder = joinPathFragments(options.name, 'src');

  await applicationGenerator(tree, options);

  const allFilesInSrc = tree
    .listChanges()
    .filter((c) => c.path.includes(srcFolder));

  for (const f of allFilesInSrc) {
    tree.delete(f.path);
  }

  const generatorFolder = joinPathFragments(__dirname, './files');
  generateFiles(tree, generatorFolder, appRoot, {
    ...options,
    snakeCase,
    pascalCase,
  });

  if (!options.db) {
    const dbFiles = tree
      .listChanges()
      .filter(
        (c) => c.path.includes('/db/') || c.path.includes('/repositories/'),
      );

    for (const c of dbFiles) tree.delete(c.path);
  }

  if (!options.i18n) {
    const i18nFiles = tree.listChanges().filter((c) => c.path.includes('i18n'));
    for (const c of i18nFiles) tree.delete(c.path);

    //   todo generate i18n languages
  }

  await formatFiles(tree);
}

export default appGenerator;
